Update Clutter-Gst to the new Core API
The ClutterMedia interface has been changed to fix bug 1404 and
bug 1403. As of commit 5a4f9c50 on clutter-core, the implementations
of ClutterMedia must only override the GObject properties defined
by the interface.

This commit also updates the examples to the API changes in
Clutter core, especially the ClutterText actor superceding the
ClutterLabel used in the audio and video-player examples. The
video-player example has also been changed to use the new Implicit
Animations API.