[videosink] Introduce ClutterGstRenderer
This structure is meant to abstract the work we have to do to actually
paint a frame. This way we can provide several backends to render the
same frame format (eg. between GLSL shaders or ARB fp1.0 fragment
programs) or build the caps of the GStreamer sink at run time after
having probed the GL implementation.

I420 and YV12 have been converted to work with this new abstraction.