[videosink] Add a I420 to RGBA fragment program
Things are a bit convoluted:
  * we want to keep GLSL shaders (eg OpenGL ES 2.0 does not have
    the GL_ARB_fragment_program extension)
  * we want to default to assembly fragment programs as they will support
    as broader range of hardware

Now, those fragment programs are written in Cg, compiled with the Cg
compiler and its arbfp1 profile and then turned into a header file
thanks to a small python script. Sounds fun isn't it?

As people may not want to install the Cg compiler, let's commit the
actual fragment program. The configure script will detect if you have
cgc and add targets to build the pso file accordingly.