Use the Cogl material API for multi-texturing rather than GL directly
Instead of manually binding the higher layer textures when rendering a
YUV image it now just sets the material layers. This is necessary now
that Clutter stores the texture filter settings in the material rather
than the Cogl texture. This meant that nothing was changing the GL
filter settings from GL_LINEAR_MIPMAP so the texture lookups would
fail.

It should also avoid potential bugs if anything else in the scene is
using multi-texturing because it won't conflict with Cogl's cache of
which texture units are enabled.

The test for the multi-texturing API has been removed and instead it
just checks that the number of layers is at least 3. The API is
available since OpenGL 1.2 and Clutter requires at it to run so it is
not necessary to check.