The shaders headers could be either in $srcdir or $builddir
The headers with the ARB fp1.0 shaders are generated by the Cg compiler.
However, it's quite common not to have that compiler (and automated
build systems are quite unlikely able to have it) so we ship these
generated headers in tarballs and commit them to the git repository.

Now we want to be able to do out of tree builds with and without the Cg
compiler. Including both $top_builddir/clutter-gst/shaders and
$top_srcdir/clutter-gst/shaders (in that order) fix all the cases.

This also fix make distcheck.