Remove conditional compilation on CLUTTER_COGL_HAS_GL
We should not need this as the direction towards the video-sink would
like to go is to detect the GL capabilities at run time. It's a good
occasion to get rid of the direct glUniform1iARB() calls and use
cogl_program_uniform_1i() instead.

Another change has sneaked in this commit: GL_MAX_TEXTURE_UNITS_ARB was
changed to GL_MAX_TEXTURE_UNITS as both GL and GL ES should have this
define.