[videosink] Remove the use-shaders property
This is an API change.

This property had been unused for a couple of months. Basically the use
of shaders depends on runtime detection.
For those who still do not want to use the colorspacing shaders, you can
use a capsfilter before the clutter sink to request RGB. See
tests/test-rgb-upload.c for an example of doing just that.