[videosink] Reword of the init/exit code path
A large part of the idle handler was devoted to initialization (in a
ugly way). Let's just abstract a bit this code and use Renderer::init()
instead. The paint signal handlers are now attached in the ::init()
functions.
Then no real exit function was called for the renderers. Fix that.