Get Cogl to flush primitives when using fragment shaders
Cogl doesn't support changing OpenGL state to modify how Cogl primitives
work, but it also doesn't support ARBfp which we currently depend on. For
now we at least ask Cogl to flush any batched primitives so we avoid
binding our shader across the wrong geometry, but there is a risk that
Cogl may start to use ARBfp internally which will conflict with us.