[debug] Add a debugging insfrastructure.
Steal the GTK+/Clutter way of doing debug categories and messages.

The CLUTTER_GST_DEBUG environment variable can contain a comma separated
list of categories. Debugging messages with then be printed if the
library has been compiled with this feature enabled.

Note that this feature is always built-in by default, --disable-debug
will disable it tough.