[VideoTexture] get_playbin() is now get_pipeline()
This is an API change.

That change was motivated by the fact the former function name is
leaking that we are using the playbin element. This is an implementation
detail we don't want in a stable API. The GstPipeline used inside
VideoTexture could be playbin, playbin2, playbin42 or even a user
supplied pipeline.