[VideoSink] Make the alpha component of the GL color play with the shaders
The idea here is to generate a premultiplied alpha texture of the video
so the user can blend it in a clutter scene and do some fancy effects.
Through the use of clutter_actor_set_opacity() on the texture, we are
now able to make a video transparent.

This is for the YUV -> RGB shaders only for now. RGB uploads + alpha
will have to wait a bit.