[VideoTexture] Fix notifications of duration
The "duration" message is sent on the GstBus when an element has a new
duration. Then, this duration has to be queried from the pipeline.

Instead of that we tried to parse the duration message that returned
GST_CLOCK_TIME_NONE (which is the way to say "we have a new duration,
please query it.") and we emitted a new signal with that value.

Let's try to limit the duration signal emission even further. The signal
will only be emitted if it's useful to update a UI (1s precision). if
one need a more accurate duration, he can always get the duration
property when needed.