[VideoSink] Use a custom GSource to signal a pending buffer
Instead of creating, every single frame, a new idle source and attach it
to the clutter GMainContext, we now attach, while the video is playing,
a GSource that will signal when there is a new buffer pending.

As a result, we do less work by frame -- which is always good -- and the
code is way more readable -- wich makes me very happy.

Based on an initial patch from Ole André Vadla Ravnås.

Fixes: OHB#1657