[Audio] Fix notifications of duration
The "duration" message is sent on the GstBus when an element has a new
duration. Then, this duration has to be queried from the pipeline.

Instead of that we tried to parse the duration message that returned
GST_CLOCK_TIME_NONE (which is the way to say "we have a new duration,
please query it.") and we emitted a new signal with that value.

This is basically the commit bd3a5266206ff3c730674655cea93a8d689efd8d
but for ClutterGstAudio.