video-texture: Add support for an "idle" material
When not playing a file (say if you create a ClutterGstVideoTexture
without a file name) you may want to control how the empty VideoTexture
should look like.

For that ClutterGstVideoTexture now lets you define an "idle" material
that will be used when not having a frame to display.

The default behaviour is to paint the texture black when there is nothing
else to show.