sink: It's no longer necessary to set a dummy shader
Since we we now use cogl_material_set_user_program to use a CoglProgram
instead of calling cogl_program_use in a pre-paint I don't think we need
to worry about interfering with the clutter shader stack.

This improves the workingness of clutter-gst when using shaders for me,
though there are still some problems to resolve.