sink: Use consistent material management
This makes all the renderer code paths use the same approach to managing
CoglMaterials. In the xyz_init callback they each create a "template"
material associating any required CoglProgram with the template and
adding n empty layers depending on how many planes are used for that
format. Whenever we upload new texture data we derive a new "paint
material" from the template, set the texture handles as layers and set
the material on the ClutterTexture to be used for painting.

This style is in line with the recommended way to manage CoglMaterials
since it avoids repeated modification of materials.