video-texture: Allow to make accurate seeks
ClutterGstVideoTexture has now a property (and its equivalent getter
and setter) to set seek flags to customize a bit the seek done by
set_progress(). For now, we have an "ACCURATE" flag that maps directly
to GST_SEEK_FLAG_ACCURATE.

To be able to export correctly this new API to g-i and bindings, we need
to generate the enum types for ClutterGstSeekFlags.

Fixes: http://bugzilla.clutter-project.org/show_bug.cgi?id=2279