video-texture: Fix an pimp the audio-volume behaviour
The one that wrote the code in the first place read the documentation a
bit too quickly. "volume" in indeed between 0. and 10.0, the 100% is 1.0
and thus we were setting the sound way over 100% most of the time.

While at it, pimped a bit the code to use the StreamVolume interface,
make sure we present a cubic volume to the user for a use with a UI and
forward the volume notification the sink (say pulsesink and someone is
playing with the slider in the application tab of GNOME's sound
preferences.