video-texture: Initial support for download buffering
This patch introduces new APIs to get and set the buffering mode. We
still default to the old behaviour and people will have to opt in to use
download buffering.

VideoTexture sorts some details for you and exposes a download-buffering
signal that is supposed to give the start and end of the current segment
of downloaded data you're in. 'start' is always 0 for now.