sink: send navigation commands on certain key events
The up, down, left and right navigation keys cannot be represented as a
single UTF-8 character and using XKeysymToString to convert the key to a
string would require clutter-gst to depend on Xlib. Therefore, use
navigation commands when handling these key events and also send the
activate command when the return key is pressed.