build: Link against -lGL when we know it's safe to
We are using a gl call (glGetIntergerv). However, detecting wich GL
headers and library to use is getting impossible (as nowedays, cogl even
has a runtime selection of which library should be dlopened, so it could
be either one of flavours of the GLES1 DSO or the GLES2 DSO, or the Big
GL one).

So, really, we should not have any direct GL call in clutter-gst. The
solution is coming, but in the mean time, just link against -lGL when
it's safe to assume so.