refactoring: Do not depend on any GL symbol
Since Cogl 1.8.0 we can use cogl_renderer_get_n_fragment_texture_units()
to know the number of texture image units available for fragment
shaders and we can thus avoid using the direct GL call glGetIntegerv and
linking against libGL.

This does not work on OS X though as we can't get the CoglContext out of
the ClutterBackend is that case. Make sure to special case this.

Fixes: https://bugzilla.gnome.org/show_bug.cgi?id=657225