player: force reset of text bit in playing flags
It turns out that the playbin2 implementation applies no changes to
the pipeline when you're setting the "suburi" or "current-text"
properties. The changes are only applied when the pipeline is
reconfigured. One way to reconfigure the pipeline is to touch the
playing flags.

This patch wrap the manipulation of these properties around a get/set
of the "flags" property to force the reconfiguration of the pipeline.