configure: Use the release status define to determine default flags
The --enable-debug option had a special m4_define to set the default
value depending on whether the minor version of the release number
suggests this is a git build. This patch changes it to use the
clutter_gst_release_status define instead so that it by default is
enabled only for git builds. The --enable-maintainer-flags option now
uses this too.

The help strings now have some added square brackets otherwise m4
doesn't seem to expand the macro properly.

https://bugzilla.gnome.org/show_bug.cgi?id=661781